Got shmem window working again... perhaps compiler problem?

+++ include/RAJA/policy/cuda/nested/Tile.hpp
@@ -35,7 +35,7 @@ struct CudaStatementExecutor<Tile<ArgumentId, TPol, seq_exec, EnclosedStmts...>,
   static
   inline
   __device__
-  void exec(Data &data, int num_logical_blocks, int logical_block)
+  void exec(Data &data, long num_logical_blocks, long logical_block)
   {
     // Get the segment referenced by this Tile statement
     auto &segment = camp::get<ArgumentId>(data.segment_tuple);
@@ -44,13 +44,13 @@ struct CudaStatementExecutor<Tile<ArgumentId, TPol, seq_exec, EnclosedStmts...>,
     using segment_t = camp::decay<decltype(segment)>;
     segment_t orig_segment = segment;
 
-    int chunk_size = TPol::chunk_size;
+    long chunk_size = TPol::chunk_size;
 
     // compute trip count
-    int len = segment.end() - segment.begin();
+    long len = segment.end() - segment.begin();
 
     // Iterate through tiles
-    for (int i = 0;i < len;i += chunk_size) {
+    for (long i = 0;i < len;i += chunk_size) {
 
       // Assign our new tiled segment
       segment = orig_segment.slice(i, chunk_size);
